+++ cli/Cargo.toml
@@ -21,6 +21,7 @@ serde_yaml = "0.9.34"
 log = "0.4"
 flate2 = "1.1.0"
 tar = "0.4.44"
+mockito = "1.7.0"
 
 [dev-dependencies]
 tempfile = "3.10.1"
