+++ cli/src/commands/mod.rs
@@ -3,4 +3,3 @@ pub mod find;
 pub mod storybook;
 pub mod update_readme;
 pub mod r#use;
-pub use r#use::extract_tar_subdir;
